# openapi.yaml
openapi: 3.0.0
info:
  title: Customer Management API
  version: 1.0.0
paths:
  /customers:
    post:
      summary: Create a new customer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: Customer created successfully
        '400':
          description: Invalid request
  /customers/{id}:
    get:
      summary: Retrieve a customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found

    put: 
      summary: Update a customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid Email
        '404':
          description: Customer not found
components:
  schemas:
    CustomerRequest:
      type: object
      properties:
        name:
          type: string
          description: Full name of the customer.
          minLength: 1
        email:
          type: string
          description: Email address of the customer.
          format: email
          minLength: 1
        annualSpend:
          type: number
          format: double
        lastPurchaseDate:
          type: string
          format: date-time
      required:
        - name
        - email
    CustomerResponse:
      allOf:
        - $ref: '#/components/schemas/CustomerRequest'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            tier:
              type: string
              enum:
                - Silver
                - Gold
                - Platinum